+++ .editorconfig
@@ -1,0 +1,3 @@
+[*]
+indent_style = space
+indent_size = 2
