+++ .vscode/settings.json
@@ -3,7 +3,6 @@
   "editor.codeActionsOnSave": {
     "source.fixAll.eslint": true
   },
-  "editor.tabSize": 2,
   "search.exclude": {
     "**/build": true,
     "package-lock.json": true
