+++ src/types/index.d.ts
@@ -1,0 +1,12 @@
+declare interface Author {
+  name: string;
+  link: string;
+}
+
+declare interface Word {
+  text: string;
+  meaning: string;
+  synonyms: string[];
+  examples: string[];
+  authors: Author[];
+}
